==> 1-DataImporter.ipynb <==
# 1 Data Importer

# packages gerais
import pandas as pd

# cities dataset
df_cities = pd.read_csv("C:/Users/yfreitas/Documents/pythonSamples/ProjectII/Files/cities.csv")
# product dataset
df_product = pd.read_csv("C:/Users/yfreitas/Documents/pythonSamples/ProjectII/Files/product.csv")
# forecast revenue dataset
df_forecast_revenue = pd.read_csv("C:/Users/yfreitas/Documents/pythonSamples/ProjectII/Files/forecast_revenue_real_data.csv")
# sales
df_sales = pd.read_csv("C:/Users/yfreitas/Documents/pythonSamples/ProjectII/Files/sales.csv")

### 1.1 Collecting the firsts basic information from the dataset

df_cities.head()

df_cities.count()

df_product.head()

df_product.count()

df_forecast_revenue.head()

df_forecast_revenue.count()

df_sales.head()

df_sales.count()

### Add in temporary files to be possible use that in the following notebooks

import joblib

joblib.dump(df_sales, 'df_sales.pkl')
joblib.dump(df_cities, 'df_cities.pkl')
joblib.dump(df_product, 'df_product.pkl')
joblib.dump(df_forecast_revenue, 'df_forecast_revenue.pkl')
