==> 2-DataUnderstanding.ipynb <==
# 1 Import from pkl files

import joblib
import matplotlib.pyplot as plt

df_sales = joblib.load('df_sales.pkl')
df_product = joblib.load('df_product.pkl')
df_cities = joblib.load('df_cities.pkl')
df_forecast_revenue = joblib.load('df_forecast_revenue.pkl')

# 2 Understanding sales info

df_sales.head(10)

df_sales.tail(10)

df_sales.info()

###### After check the data type columns above, we will need to convert some of them. Among them, initially we will need to cast the field date to DATE type.

df_sales[(df_sales['promo_discount_type_2'].notna()) & (df_sales['promo_discount_type_2'] != '')]

df_sales['promo_bin_2'].unique()

df_sales['promo_bin_1'].unique()

df_sales['promo_type_1'].unique()

df_sales[(df_sales['promo_discount_2'].notna()) & (df_sales['promo_discount_2'] != '')]

df_sales[(df_sales['sales'].isna()) | (df_sales['sales'] == '')]

df_sales[(df_sales['price'].isna()) | (df_sales['price'] == '')]

df_sales[(df_sales['product_id'] == 'P0664') & (df_sales['store_id']=='S0002')].tail(30)


# Counting the number of sales by product_id
sales_count_by_product = df_sales.groupby('product_id').size().reset_index(name='sales_count')

# Sorting by the biggest count
sales_count_by_product = sales_count_by_product.sort_values(by='sales_count', ascending=False)
sales_count_by_product

df_sales['date'].max()

df_sales['date'].min()

###### Another necessary task will be to replace NaN values with something meaningful. For columns of type string, replace NaN values with a meaningful string. For columns of type float, replace NaN values with a meaningful float. Additionally, for all titles without a name, replace them with the field meaning. We could use the following logic to replace NaN values:-For string columns: Replace NaN with a specific string, such as "Unknown" or "No Data". -For numeric columns: Replace NaN with a specific value, such as 0 or the mean/median of the column. -For title fields: Replace NaN or missing values with the corresponding field name or another relevant default value.
###### Unamed 0 title by ID
###### store_id (all records are fulfilled)
###### product_id (all records are fulfilled)
###### date (all records are fulfilled)
###### sales by 0.0 (because the type is float in this case)
###### revenue by 0.0 (because the type is float in this case)
###### stock by 0.0 (because the type is float in this case)
###### price by 0.0 (because the type is float in this case)
###### promo_type_1 (all records are fulfilled)
###### promo_bin_1 by 'No promo bin 1'
###### promo_type_2 by 'No promo type 2'
###### promo_bin_2 by 'No promo bin 2'
###### promo_discount_2 by 0.0 (because the type is float in this case)
###### promo_discount_type_2 by 'No promo discount type 2'

# 3 Understanding cities info

df_cities.head(10)

df_cities['country_id'].unique()

df_cities.info()

# Check for duplicates in the 'stored_id' column
df_cities[df_cities['store_id'].duplicated()]

# 4 Understanding product info

df_product.info()

df_product.head(10)

df_product.tail(10)

df_product['cluster_id'].unique()

###### We need to following a similar approach than we did above.
###### product_length by median (because the type is float and we have only 19 cases with NaN)
###### product_depth by median (because the type is float and we have only 17 cases with NaN)
###### product_width by median (because the type is float and we have only 17 cases with NaN)
###### cluster_0 by 'No cluster'

# 5 Understanding forecast revenue info

df_forecast_revenue.info()

df_forecast_revenue.head(12)

df_forecast_revenue.tail(15)

###### After check the data type columns above, we will need to convert date field. Initially we will need to cast the field date to DATE type.
