==> 6 -Compare results.ipynb <==
import pandas as pd

# loading csv files
metrics_2019 = pd.read_csv('2019_mse_results.csv')
metrics_all_arimax = pd.read_csv('metrics_results_arimax.csv')

# Calculate means Metrics DataFrame metrics_all
mean_metrics_2019 = metrics_2019[['mse', 'rmse', 'mae']].mean()

# Calculate Metrics DataFrame metrics_2019_arimax
mean_metrics_2019_arimax = metrics_all_arimax[['mse', 'rmse', 'mae']].mean()

# Imprimir os resultados
print("Means ARIMA with dataset 2019 information:")
print(mean_metrics_2019)
print("\nMeans ARIMAX with all information we have:")
print(mean_metrics_2019_arimax)

 # CONCLUSION
 ### This ARIMA model performs better with the 2019 dataset, while the ARIMAX model performs better when using all the data we have.
